Small code corrections
